Restrict certain board operations to assigned users or board author

+++ controllers/boards.controller.ts
@@ -2,6 +2,7 @@ import { NextFunction, Request, Response } from 'express';
 import prisma from '../prisma/prisma';
 import createBoardDTO from '../validators/boards/create-board.dto';
 import UpdateBoardTitleDTO from '../validators/boards/update-board-title';
+import UserIdSchema from '../validators/validateUserId';
 
 const BoardsController = {
   // [TODO - delete this endpoint for production]
@@ -15,11 +16,29 @@ const BoardsController = {
     }
   },
 
-  // [TODO - add authorization to allow getting board by ID only for author or assigned users]
+  // getting board by ID is allowed only for  assigned users
   getById: async (req: Request, res: Response, next: NextFunction) => {
     const { boardId } = req.params;
+    let requestAuthor;
+    try {
+      requestAuthor = UserIdSchema.parse(req.body);
+    } catch (error) {
+      return res.status(400).json({ error: 'Invalid user data' });
+    }
 
     try {
+      // check if user who made the request is assigned to the board
+      const assignedUser = await prisma.userOnBoard.findUnique({
+        where: {
+          userId_boardId: {
+            userId: requestAuthor.userId,
+            boardId,
+          },
+        },
+      });
+
+      if (!assignedUser) return res.status(403).json({ error: 'Access only for assigned users!' });
+
       const board = await prisma.board.findUnique({
         where: { id: boardId },
         include: {
@@ -83,14 +102,17 @@ const BoardsController = {
     }
   },
 
-  // [TODO - add authorization to allow editing board only by the author of the board]
+  // editing board title available only for the board's author
   editBoardTitle: async (req: Request, res: Response, next: NextFunction) => {
     const { boardId } = req.params;
+
     let title;
+    let requestAuthor;
 
     try {
       const bodyObj = UpdateBoardTitleDTO.parse(req.body);
       title = bodyObj.title;
+      requestAuthor = bodyObj.userId;
     } catch (error) {
       return res.status(400).json({ error: 'Invalid data' });
     }
@@ -101,7 +123,10 @@ const BoardsController = {
           id: boardId,
         },
       });
+
       if (!board) return res.status(404).json({ error: 'Board not found...' });
+      if (board.authorId !== requestAuthor)
+        return res.status(403).json({ error: 'Editing the board is only allowed by the author' });
 
       await prisma.board.update({
         where: {
@@ -118,15 +143,26 @@ const BoardsController = {
     }
   },
 
-  // [TODO - add authorization to allow adding users to the board only by the author of the board]
+  // adding users to the board allowed only for the author of the board
   addUserToBoard: async (req: Request, res: Response, next: NextFunction) => {
     const { boardId, userId } = req.params;
+
+    let requestAuthor;
+    try {
+      requestAuthor = UserIdSchema.parse(req.body);
+    } catch (error) {
+      return res.status(400).json({ error: 'Invalid user data' });
+    }
 
     try {
       const board = await prisma.board.findUnique({
         where: { id: boardId },
       });
       if (!board) return res.status(404).json({ error: 'Board not found...' });
+      if (board.authorId !== requestAuthor.userId)
+        return res
+          .status(403)
+          .json({ error: 'Adding users only allowed by the creator of the board.' });
 
       const user = await prisma.user.findUnique({
         where: { id: userId },
@@ -166,21 +202,25 @@ const BoardsController = {
   // [TODO - add authorization to allow deleting a board only by the author of the board]
   deleteBoard: async (req: Request, res: Response, next: NextFunction) => {
     const { boardId } = req.params;
+    let requestAuthor;
+
+    try {
+      requestAuthor = UserIdSchema.parse(req.body);
+    } catch (error) {
+      return res.status(400).json({ error: 'Invalid user data' });
+    }
 
     try {
       const board = await prisma.board.findUnique({
         where: {
           id: boardId,
         },
-        select: {
-          author: {
-            select: {
-              id: true, // author id needed for future authorization
-            },
-          },
-        },
       });
       if (!board) return res.status(404).json({ error: 'Board not found...' });
+      if (board.authorId !== requestAuthor.userId)
+        return res
+          .status(403)
+          .json({ error: "Deleting the board is only allowed by it's author." });
 
       await prisma.board.delete({
         where: {
@@ -197,13 +237,24 @@ const BoardsController = {
   // [TODO - add authorization to allow deleting a board only by the author of the board]
   deleteUserFromBoard: async (req: Request, res: Response, next: NextFunction) => {
     const { boardId, userId } = req.params;
+    let requestAuthor;
 
     try {
-      // Check if the board exists
+      requestAuthor = UserIdSchema.parse(req.body);
+    } catch (error) {
+      return res.status(400).json({ error: 'Invalid user data' });
+    }
+
+    try {
+      // Check if the board exists and user authorization
       const board = await prisma.board.findUnique({
         where: { id: boardId },
       });
       if (!board) return res.status(404).json({ error: 'Board not found...' });
+      if (board.authorId !== requestAuthor.userId)
+        return res
+          .status(403)
+          .json({ error: "Removing users from the board is only allowed by it's author." });
 
       // Check if the user exists
       const user = await prisma.user.findUnique({
